+++ clx_task_management/models/__init__.py
@@ -8,3 +8,4 @@ from . import request_form
 from . import project
 from . import sale_order_line
 from . import sale_order
+from . import res_config_settings